Format and Print Coords or Waypoints

Format and print objects of class "coords" or "waypoints".

Usage

# S3 method for class 'coords'
print(x, ..., max = NULL)

# S3 method for class 'waypoints'
print(x, ..., max = NULL)

# S3 method for class 'coords'
format(x, ..., usenames = TRUE)

# S3 method for class 'waypoints'
format(x, ..., usenames = TRUE)

ll_headers(aswidth, fmt)

Arguments

x

object of class "coords" created by function as_coords(), or class "waypoints" created by function as_waypoints().

...

further arguments passed to or from other methods.

max

numeric or NULL, specifying the maximal number of entries to be printed. By default, when NULL, getOption("max.print") used.

usenames

logical, whether or not to include names in formatted output; default TRUE.

aswidth

character vector, used to match width of headers to formatted output.

fmt

integer, 1L, 2L or 3L, specifying the required coordinate format.

Value

The format() methods for both classes "coords" and "waypoints" return a character vector, respectively of length length(x) or nrow(x), and containing values formatted in decimal degrees, degrees and minutes, or degrees, minutes and seconds as appropriate.

Details

The format() methods for "coords" and "waypoints" objects output elegantly formatted character vector representations of their arguments, which are used by their respective print() methods.

Objects of class "coords" specified in degrees and minutes or in degrees, minutes and seconds and with a "latlon" attribute, and similarly specified "waypoints" objects are formatted with individual coordinate values followed by a capital letter representing the cardinal direction i.e., N, E, S or W. "coords" objects lacking a "latlon" attribute have formatted values followed by two possible cardinal directions in parentheses i.e., (N/E) for positive values and (S/W) for negative values. Values of "coords" or "waypoints" objects in decimal degrees are formatted prefixed with their sign, if negative; cardinal direction is not shown, but for "coords" objects with a "latlon" attribute, the formatted values are suffixed by either lat or lon.

Prior to formatting and printing, "coords" or "waypoints" objects are checked to ensure that their "valid" attribute (in the case of a "coords" object), or "validlat" and "validlon" attributes (in the case of a "waypoints" object) are present and all TRUE i.e., valid. If these attributes are found to contain any FALSE i.e. invalid values, a warning is issued and similarly, if these attributes are missing, a warning is issued and the objects are re-validated as described under validate().

ll_headers() outputs the headings "Latitude ... Longitude" formatted to the same width as argument aswidth, adjusted for format fmt and is primarily intended for use by the print() method for class "waypoints".

See also

Examples

## Continuing example from `as_coords()`...

## Print named "coords" object in degrees and minutes,
## 'silently' using S3 print() method
dm
#> Nelson's Column                           51°0000030′ N
#> Ostravice                                 49°0000032′ N
#> Tally Ho                                  48°0000006′ N
#> Washington Monument                       38°0000053′ N
#> Null Island                                0°0000000′ N
#> Tristan da Cunha                          37°0000006′ S
#> Mawson Peak                               53°0000006′ S
#> Silvio Pettirossi International Airport   25°0000014′ S
#> Nelson's Column                            0°0000007′ W
#> Ostravice                                 18°0000023′ E
#> Tally Ho                                 122°0000046′ W
#> Washington Monument                       77°0000002′ W
#> Null Island                                0°0000000′ E
#> Tristan da Cunha                          12°0000017′ W
#> Mawson Peak                               73°0000031′ E
#> Silvio Pettirossi International Airport   57°0000031′ W

## Print explicitly using S3 print() method, specifying
## the maximal number of entries to be printed
print(dm, max = 14)
#> Nelson's Column       51°0000030′ N
#> Ostravice             49°0000032′ N
#> Tally Ho              48°0000006′ N
#> Washington Monument   38°0000053′ N
#> Null Island            0°0000000′ N
#> Tristan da Cunha      37°0000006′ S
#> Mawson Peak           53°0000006′ S
#>  [ reached 'max' / getOption("max.print") -- omitted 9 entries ]

## Format as a fixed-width character vector,
## with names...
format(dm)
#>  [1] "Nelson's Column                           51°0000030′ N"
#>  [2] "Ostravice                                 49°0000032′ N"
#>  [3] "Tally Ho                                  48°0000006′ N"
#>  [4] "Washington Monument                       38°0000053′ N"
#>  [5] "Null Island                                0°0000000′ N"
#>  [6] "Tristan da Cunha                          37°0000006′ S"
#>  [7] "Mawson Peak                               53°0000006′ S"
#>  [8] "Silvio Pettirossi International Airport   25°0000014′ S"
#>  [9] "Nelson's Column                            0°0000007′ W"
#> [10] "Ostravice                                 18°0000023′ E"
#> [11] "Tally Ho                                 122°0000046′ W"
#> [12] "Washington Monument                       77°0000002′ W"
#> [13] "Null Island                                0°0000000′ E"
#> [14] "Tristan da Cunha                          12°0000017′ W"
#> [15] "Mawson Peak                               73°0000031′ E"
#> [16] "Silvio Pettirossi International Airport   57°0000031′ W"

## ...or without them
format(dm, usenames = FALSE)
#>  [1] " 51°0000030′ N" " 49°0000032′ N" " 48°0000006′ N" " 38°0000053′ N"
#>  [5] "  0°0000000′ N" " 37°0000006′ S" " 53°0000006′ S" " 25°0000014′ S"
#>  [9] "  0°0000007′ W" " 18°0000023′ E" "122°0000046′ W" " 77°0000002′ W"
#> [13] "  0°0000000′ E" " 12°0000017′ W" " 73°0000031′ E" " 57°0000031′ W"

###
## Continuing example from `as_waypoints()`...

## Print named "waypoints" object in degrees and minutes,
## 'silently' using S3 print() method
wp
#>                                                Latitude        Longitude 
#>                                           ______________  _______________
#> Nelson's Column                           51°30′00027″ N    0°07′00040″ W
#> Ostravice                                 49°32′00046″ N   18°23′00054″ E
#> Tally Ho                                  48°06′00026″ N  122°46′00043″ W
#> Washington Monument                       38°53′00022″ N   77°02′00006″ W
#> Null Island                                0°00′00000″ N    0°00′00000″ E
#> Tristan da Cunha                          37°06′00042″ S   12°17′00019″ W
#> Mawson Peak                               53°06′00017″ S   73°31′00002″ E
#> Silvio Pettirossi International Airport   25°14′00024″ S   57°31′00009″ W

## Print explicitly using S3 print() method, specifying
## the maximal number of entries to be printed
print(wp, max = 21)
#>                            Latitude        Longitude 
#>                       ______________  _______________
#> Nelson's Column       51°30′00027″ N    0°07′00040″ W
#> Ostravice             49°32′00046″ N   18°23′00054″ E
#> Tally Ho              48°06′00026″ N  122°46′00043″ W
#> Washington Monument   38°53′00022″ N   77°02′00006″ W
#> Null Island            0°00′00000″ N    0°00′00000″ E
#> Tristan da Cunha      37°06′00042″ S   12°17′00019″ W
#> Mawson Peak           53°06′00017″ S   73°31′00002″ E
#>  [ reached 'max' / getOption("max.print") -- omitted 1 rows ]

## Format as a fixed-width character vector,
## with names...
format(wp)
#> [1] "Nelson's Column                           51°30′00027″ N    0°07′00040″ W"
#> [2] "Ostravice                                 49°32′00046″ N   18°23′00054″ E"
#> [3] "Tally Ho                                  48°06′00026″ N  122°46′00043″ W"
#> [4] "Washington Monument                       38°53′00022″ N   77°02′00006″ W"
#> [5] "Null Island                                0°00′00000″ N    0°00′00000″ E"
#> [6] "Tristan da Cunha                          37°06′00042″ S   12°17′00019″ W"
#> [7] "Mawson Peak                               53°06′00017″ S   73°31′00002″ E"
#> [8] "Silvio Pettirossi International Airport   25°14′00024″ S   57°31′00009″ W"

## ...or without them
format(wp, usenames = FALSE)
#> [1] " 51°30′00027″ N    0°07′00040″ W" " 49°32′00046″ N   18°23′00054″ E"
#> [3] " 48°06′00026″ N  122°46′00043″ W" " 38°53′00022″ N   77°02′00006″ W"
#> [5] "  0°00′00000″ N    0°00′00000″ E" " 37°06′00042″ S   12°17′00019″ W"
#> [7] " 53°06′00017″ S   73°31′00002″ E" " 25°14′00024″ S   57°31′00009″ W"

rm(dm, wp)