Latitude or Longitude Attribute for Coords

latlon()<- adds the attribute "latlon" to objects of class "coords", or modifies an existing "latlon" attribute.

Usage

latlon(cd) <- value

Arguments

cd

object of class "coords".

value

a boolean vector of length 1L or of length(cd).

Value

Argument cd is returned with boolean vector attribute "latlon" updated as appropriate.

Details

Attribute "latlon" is a boolean vector of length 1L or length(cd) for which TRUE values represent latitude and FALSE values represent longitude. Setting this attribute to any other length will result in an error. A boolean vector of length 1L signifies that values are all latitude if TRUE, or longitude if FALSE.

This attribute is used in formatting printed output and also by validate(). Indeed, the values of cd are revalidated every time attribute "latlon" is added or changed.

Examples

## Continuing example from `coords()`, named numeric vector representing degrees and minutes

## Create "coords" object of degrees and minutes
coords(num_dm) <- 2

## Set "latlon" attribute to FALSE, length 1
latlon(num_dm) <- FALSE
num_dm
#>  51°30.4659' E  Nelson's Column
#>  49°32.7726' E  Ostravice
#>  48°06.4339' E  Tally Ho
#>  38°53.3696' E  Washington Monument
#>   0°00.0000' E  Null Island
#>  37°06.7044' W  Tristan da Cunha
#>  53°06.2869' W  Mawson Peak
#>  57°31.1536' W  Silvio Pettirossi International Airport
#>   0°07.6754' W  Nelson's Column
#>  18°23.9137' E  Ostravice
#> 122°46.7203' W  Tally Ho
#>  77°02.1145' W  Washington Monument
#>   0°00.0000' E  Null Island
#>  12°17.3178' W  Tristan da Cunha
#>  73°31.0370' E  Mawson Peak
#>  25°14.4093' W  Silvio Pettirossi International Airport

## Set "latlon" attribute to TRUE and FALSE (each n=8)
latlon(num_dm) <- rep(c(TRUE, FALSE), each = 8)
num_dm
#>  51°30.4659' N  Nelson's Column
#>  49°32.7726' N  Ostravice
#>  48°06.4339' N  Tally Ho
#>  38°53.3696' N  Washington Monument
#>   0°00.0000' N  Null Island
#>  37°06.7044' S  Tristan da Cunha
#>  53°06.2869' S  Mawson Peak
#>  57°31.1536' S  Silvio Pettirossi International Airport
#>   0°07.6754' W  Nelson's Column
#>  18°23.9137' E  Ostravice
#> 122°46.7203' W  Tally Ho
#>  77°02.1145' W  Washington Monument
#>   0°00.0000' E  Null Island
#>  12°17.3178' W  Tristan da Cunha
#>  73°31.0370' E  Mawson Peak
#>  25°14.4093' W  Silvio Pettirossi International Airport

## Reversing latitude and longitude results in an implausible 
## longitude value and a warning message 
latlon(num_dm) <- rep(c(FALSE, TRUE), each = 8)
#> Warning: Validation failed!
num_dm
#> Warning: Printing invalid coords!
#>  51°30.4659' E  Nelson's Column
#>  49°32.7726' E  Ostravice
#>  48°06.4339' E  Tally Ho
#>  38°53.3696' E  Washington Monument
#>   0°00.0000' E  Null Island
#>  37°06.7044' W  Tristan da Cunha
#>  53°06.2869' W  Mawson Peak
#>  57°31.1536' W  Silvio Pettirossi International Airport
#>   0°07.6754' S  Nelson's Column
#>  18°23.9137' N  Ostravice
#> 122°46.7203' S  Tally Ho
#>  77°02.1145' S  Washington Monument
#>   0°00.0000' N  Null Island
#>  12°17.3178' S  Tristan da Cunha
#>  73°31.0370' N  Mawson Peak
#>  25°14.4093' S  Silvio Pettirossi International Airport

rm(num_dm)