Review Coordinates and Waypoints Validity

review() review validity of elements of "coords" and "waypoints" objects.

Usage

review(x, ...)

# S3 method for class 'coords'
review(x, ..., show_n = 20L)

# S3 method for class 'waypoints'
review(x, ..., show_n = 20L)

Arguments

x

object of class "coords" or "waypoints".

...

further arguments passed to or from other methods.

show_n

integer, the maximum number of invalid elements of argument x to include in the output; default 20L.

Value

The review() method for class "coords" returns a list comprising the following elements: -

allvalid

logical, whether or not all the elements of argument x are valid.

n_invalid

integer, the number of invalid elements in argument x, if any.

invalids

numeric vector including invalid elements of argument x, if any.

which_invalid

integer vector specifying which elements of argument x are invalid, if any.

The method for class "waypoints" returns a list of two sub-lists, each sub-list with elements as described for the method for class "coords", one each for latitude and longitude.

Details

review() reveals elements of "coords" and "waypoints" objects that do not conform to the criteria checked by validate(), i.e. are not valid geographic locations.

See also

Examples

## Continuing example from `validate()`, named numeric vector representing degrees and minutes,
## the erroneous first value having more than 60 minutes

## Create "coords" object of degrees and minutes
as_coords(dm, fmt = 2)
#> Nelson's Column                           51°60.4659′ (N/E)
#> Ostravice                                 49°32.7726′ (N/E)
#> Tally Ho                                  48°06.4339′ (N/E)
#> Washington Monument                       38°53.3696′ (N/E)
#> Null Island                                0°00.0000′ (N/E)
#> Tristan da Cunha                          37°06.7044′ (S/W)
#> Mawson Peak                               53°06.2869′ (S/W)
#> Silvio Pettirossi International Airport   25°14.4093′ (S/W)
#> Nelson's Column                            0°07.6754′ (S/W)
#> Ostravice                                 18°23.9137′ (N/E)
#> Tally Ho                                 122°46.7203′ (S/W)
#> Washington Monument                       77°02.1145′ (S/W)
#> Null Island                                0°00.0000′ (N/E)
#> Tristan da Cunha                          12°17.3178′ (S/W)
#> Mawson Peak                               73°31.0370′ (N/E)
#> Silvio Pettirossi International Airport   57°31.1536′ (S/W)

review(dm)
#> $allvalid
#> [1] FALSE
#> 
#> $n_invalid
#> [1] 1
#> 
#> $invalids
#> Nelson's Column   51°60.4659′ (N/E)
#> 
#> $which_invalid
#> [1] 1
#> 

###
## Continuing example from `validate()`, data frame representing waypoint names and latitude
## and longitude values in decimal degrees, the erroneous penultimate latitude having more than
## 90 degrees absolute value

## Create "waypoints" object of decimal degrees (fmt = 1)
as_waypoints(wp1, fmt = 1)
#>                                             Latitude     Longitude 
#>                                           ___________  ____________
#> Nelson's Column                            51.507765°    -0.127924°
#> Ostravice                                  49.546210°    18.398562°
#> Tally Ho                                   48.107232°  -122.778671°
#> Washington Monument                        38.889494°   -77.035242°
#> Null Island                                 0.000000°     0.000000°
#> Tristan da Cunha                          -37.111740°   -12.288630°
#> Mawson Peak                               -93.104781°    73.517283°
#> Silvio Pettirossi International Airport   -25.240156°   -57.519227°

review(wp1)
#> $Lat
#> $Lat$allvalid
#> [1] FALSE
#> 
#> $Lat$n_invalid
#> [1] 1
#> 
#> $Lat$invalids
#>  -93.104781° lat
#> 
#> $Lat$which_invalid
#> [1] 7
#> 
#> 
#> $Lon
#> $Lon$allvalid
#> [1] TRUE
#> 
#> $Lon$n_invalid
#> [1] 0
#> 
#> $Lon$invalids
#> [1] NA
#> 
#> $Lon$which_invalid
#> [1] NA
#> 
#> 

rm(dm, wp1)