Geographic or GPS Waypoint Class

waypoints() creates a robust representation of a series of geographic or GPS waypoints instantiated as an object of class "waypoints".

waypoints() and replacement form waypoints()<- also convert the format of existing objects of class "waypoints" between (i) decimal degrees, (ii) degrees and minutes, and (iii) degrees, minutes and seconds.

Usage

# S3 method for class 'waypoints'
print(x, ...)

waypoints(df, fmt = 1L)

waypoints(df) <- value

# S3 method for class 'waypoints'
format(x, usenames = TRUE, ...)

Arguments

x

an object of class "waypoints" created by function waypoints().

...

further arguments passed to or from other methods.

df

a data frame with each row representing a waypoint, comprising at least two numeric columns containing values of latitude and longitude, and optionally a character column of waypoint names (see Details).

fmt, value

an integer of value 1L, 2L or 3L, indicating the current or desired coordinate format (see Details); default 1L.

usenames

logical, whether or not to include waypoint names in formatted output.

Value

An object of classes "waypoints" and "data.frame", comprising the original data frame argument df, with latitude and longitude values possibly converted as appropriate and additional attributes: –

"fmt"

the coordinate format.

"namescol"

the position of waypoint names, if present within df.

"llcols"

the position of latitude and longitude columns within df.

"validlat" and "validlon"

logical vectors indicating whether individual latitude and longitude values are valid geographic locations.

Details

By default, the names of the waypoints should be included in a "Name" column of data frame argument df, and the latitude and longitude in the two columns immediately on the right hand side of "Name". An alternative column for waypoint names may be specified by setting an integer attribute, "namescol" indicating its position in df, while setting this attribute to NA supresses printing of waypoint names. If df has neither a "Name" column nor a "namescol" attribute, the "row.names" attribute is used for waypoint names if present in df. Similarly, alternative columns for the latitude and longitude may be specified by setting "llcols" as a length 2 integer vector attribute indicating their positions in df.

Individual values provided in the numeric vector latitude and longitude columns of data frame argument df should have a decimal point after the number of whole degrees in the case of decimal degrees, after the number of whole minutes in the case of degrees and minutes, and after the number of whole seconds in the case of degrees, minutes and seconds.

The fmt argument should be 1L to represent decimal degrees, 2L for degrees and minutes, and 3L for degrees, minutes and seconds and is used to provide both the format of values in data frame argument df to be converted into a "waypoints" object and the desired format if a "waypoints" object is to be converted to a new format. Note that on conversion of a "waypoints" object, the original data frame argument df is modified such that the latitude and longitude values are as described in the previous paragraph, and may be inspected using standard R code, see examples.

The latitude and longitude values of a newly created "waypoints" object are checked to ensure they are valid geographic locations as described under validate(). Likewise, a check is made to ensure that an existing "waypoints" object to be converted to a new format has already been validated; if not, it is re-validated.

See also

Examples

## Dataframe representing waypoint names, and latitude and longitude values
## of degrees, minutes and seconds
wp1 <- data.frame(
    name = c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
             "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"),
    lat = c(513027.95, 493246.36, 480626.04, 385322.18, 0, -370642.26, -530617.21, -251424.56),
    lon = c(-00740.53, 182354.82, -1224643.22, -770206.87, 0, -121719.07, 733102.22, -573109.21)
)

## Create "waypoints" object of degrees, minutes and seconds (fmt = 3)
waypoints(wp1, 3)
#>                         Nelson's Column   51°30′27.95″ N    0°07′40.53″ W
#>                               Ostravice   49°32′46.36″ N   18°23′54.82″ E
#>                                Tally Ho   48°06′26.04″ N  122°46′43.22″ W
#>                     Washington Monument   38°53′22.18″ N   77°02′06.87″ W
#>                             Null Island    0°00′00.00″ N    0°00′00.00″ E
#>                        Tristan da Cunha   37°06′42.26″ S   12°17′19.07″ W
#>                             Mawson Peak   53°06′17.21″ S   73°31′02.22″ E
#> Silvio Pettirossi International Airport   25°14′24.56″ S   57°31′09.21″ W

## Convert to degrees and minutes (fmt = 2)
waypoints(wp1) <- 2
wp1
#>                         Nelson's Column   51°30.4658′ N    0°07.6755′ W
#>                               Ostravice   49°32.7727′ N   18°23.9137′ E
#>                                Tally Ho   48°06.4340′ N  122°46.7203′ W
#>                     Washington Monument   38°53.3697′ N   77°02.1145′ W
#>                             Null Island    0°00.0000′ N    0°00.0000′ E
#>                        Tristan da Cunha   37°06.7043′ S   12°17.3178′ W
#>                             Mawson Peak   53°06.2868′ S   73°31.0370′ E
#> Silvio Pettirossi International Airport   25°14.4093′ S   57°31.1535′ W

## Convert to decimal degrees (fmt = 1)
waypoints(wp1) <- 1
wp1
#>                         Nelson's Column    51.507764°    -0.127925°
#>                               Ostravice    49.546211°    18.398561°
#>                                Tally Ho    48.107233°  -122.778672°
#>                     Washington Monument    38.889494°   -77.035242°
#>                             Null Island     0.000000°     0.000000°
#>                        Tristan da Cunha   -37.111739°   -12.288631°
#>                             Mawson Peak   -53.104781°    73.517283°
#> Silvio Pettirossi International Airport   -25.240156°   -57.519225°

###
## Dataframe representing unnamed latitude and longitude values in decimal degrees
wp2 <- data.frame(
    lat = c(51.507765, 49.54621, 48.107232, 38.889494, 0, -37.11174, -53.104781, -25.240156),
    lon = c(-0.127924, 18.398562, -122.778671, -77.035242, 0, -12.28863, 73.517283, -57.519227)
)

## Create "waypoints" object of decimal degrees (default fmt = 1)
waypoints(wp2)
#> 1    51.507765°    -0.127924°
#> 2    49.546210°    18.398562°
#> 3    48.107232°  -122.778671°
#> 4    38.889494°   -77.035242°
#> 5     0.000000°     0.000000°
#> 6   -37.111740°   -12.288630°
#> 7   -53.104781°    73.517283°
#> 8   -25.240156°   -57.519227°

## Add row.names
row.names(wp2) <-
    c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
      "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport")
wp2
#>                         Nelson's Column    51.507765°    -0.127924°
#>                               Ostravice    49.546210°    18.398562°
#>                                Tally Ho    48.107232°  -122.778671°
#>                     Washington Monument    38.889494°   -77.035242°
#>                             Null Island     0.000000°     0.000000°
#>                        Tristan da Cunha   -37.111740°   -12.288630°
#>                             Mawson Peak   -53.104781°    73.517283°
#> Silvio Pettirossi International Airport   -25.240156°   -57.519227°

## Convert to degrees and minutes (fmt = 2)
waypoints(wp2) <- 2
wp2
#>                         Nelson's Column   51°30.4659′ N    0°07.6754′ W
#>                               Ostravice   49°32.7726′ N   18°23.9137′ E
#>                                Tally Ho   48°06.4339′ N  122°46.7203′ W
#>                     Washington Monument   38°53.3696′ N   77°02.1145′ W
#>                             Null Island    0°00.0000′ N    0°00.0000′ E
#>                        Tristan da Cunha   37°06.7044′ S   12°17.3178′ W
#>                             Mawson Peak   53°06.2869′ S   73°31.0370′ E
#> Silvio Pettirossi International Airport   25°14.4094′ S   57°31.1536′ W

## Convert to degrees, minutes and seconds (fmt = 3)
waypoints(wp2) <- 3
wp2
#>                         Nelson's Column   51°30′27.95″ N    0°07′40.53″ W
#>                               Ostravice   49°32′46.36″ N   18°23′54.82″ E
#>                                Tally Ho   48°06′26.04″ N  122°46′43.22″ W
#>                     Washington Monument   38°53′22.18″ N   77°02′06.87″ W
#>                             Null Island    0°00′00.00″ N    0°00′00.00″ E
#>                        Tristan da Cunha   37°06′42.26″ S   12°17′19.07″ W
#>                             Mawson Peak   53°06′17.21″ S   73°31′02.22″ E
#> Silvio Pettirossi International Airport   25°14′24.56″ S   57°31′09.22″ W

## Degrees, minutes and seconds values as an ordinary R data frame
as.data.frame(wp2)
#>                                               lat           lon
#> Nelson's Column                          513028.0     -740.5264
#> Ostravice                                493246.4   182354.8232
#> Tally Ho                                 480626.0 -1224643.2156
#> Washington Monument                      385322.2  -770206.8712
#> Null Island                                   0.0        0.0000
#> Tristan da Cunha                        -370642.3  -121719.0680
#> Mawson Peak                             -530617.2   733102.2188
#> Silvio Pettirossi International Airport -251424.6  -573109.2172

## Convert to decimal degrees, format as a character vector…
waypoints(wp2) <- 1
(wp2_chr <- format(wp2))
#> [1] "                        Nelson's Column    51.507765°    -0.127924°"
#> [2] "                              Ostravice    49.546210°    18.398562°"
#> [3] "                               Tally Ho    48.107232°  -122.778671°"
#> [4] "                    Washington Monument    38.889494°   -77.035242°"
#> [5] "                            Null Island     0.000000°     0.000000°"
#> [6] "                       Tristan da Cunha   -37.111740°   -12.288630°"
#> [7] "                            Mawson Peak   -53.104781°    73.517283°"
#> [8] "Silvio Pettirossi International Airport   -25.240156°   -57.519227°"

## …and output using {base} cat()
cat(wp2_chr, fill = 26, labels = paste0("{#", 1:8, "}:"))
#> {#1}:                         Nelson's Column    51.507765°    -0.127924° 
#> {#2}:                               Ostravice    49.546210°    18.398562° 
#> {#3}:                                Tally Ho    48.107232°  -122.778671° 
#> {#4}:                     Washington Monument    38.889494°   -77.035242° 
#> {#5}:                             Null Island     0.000000°     0.000000° 
#> {#6}:                        Tristan da Cunha   -37.111740°   -12.288630° 
#> {#7}:                             Mawson Peak   -53.104781°    73.517283° 
#> {#8}: Silvio Pettirossi International Airport   -25.240156°   -57.519227°

rm(wp1, wp2)