Convert the Format of "coords" and "waypoints" Objects

Converts the format of objects of class "coords" or "waypoints" between (i) decimal degrees, (ii) degrees and minutes, and (iii) degrees, minutes and seconds.

Usage

convert(x, ...)

# S3 method for class 'coords'
convert(x, fmt, ...)

# S3 method for class 'waypoints'
convert(x, fmt, ...)

Arguments

x

object of class "coords" created by function as_coords(), or of class "waypoints" created by function as_waypoints().

...

further arguments passed to or from other methods.

fmt

integer, 1L, 2L or 3L, specifyinig the required coordinate format.

Value

The original argument x, an object of class "coords" or "waypoints" with values converted as described under details and a revised "fmt" attribute reflecting the new format.

Details

The fmt argument should be 1L to convert to decimal degrees, 2L, to convert to degrees and minutes, and 3L to convert to degrees, minutes and seconds. On conversion of a "coords" object, the original argument x is modified to have a decimal point after the number of whole degrees in the case of decimal degrees, after the number of whole minutes in the case of degrees and minutes, and after the number of whole seconds in the case of degrees, minutes and seconds.

Prior to conversion, the "coords" or "waypoints" object to be converted is checked to ensure its values represent valid geographic locations as described under validate().

See also

Examples

## Continuing example from `coords()`...
## Named "coords" object in degrees and minutes

as_coords(dm, fmt = 2)
#> Nelson's Column                           51°30.4659′ (N/E)
#> Ostravice                                 49°32.7726′ (N/E)
#> Tally Ho                                  48°06.4339′ (N/E)
#> Washington Monument                       38°53.3696′ (N/E)
#> Null Island                                0°00.0000′ (N/E)
#> Tristan da Cunha                          37°06.7044′ (S/W)
#> Mawson Peak                               53°06.2869′ (S/W)
#> Silvio Pettirossi International Airport   25°14.4093′ (S/W)
#> Nelson's Column                            0°07.6754′ (S/W)
#> Ostravice                                 18°23.9137′ (N/E)
#> Tally Ho                                 122°46.7203′ (S/W)
#> Washington Monument                       77°02.1145′ (S/W)
#> Null Island                                0°00.0000′ (N/E)
#> Tristan da Cunha                          12°17.3178′ (S/W)
#> Mawson Peak                               73°31.0370′ (N/E)
#> Silvio Pettirossi International Airport   57°31.1536′ (S/W)

## Convert to degrees, minutes and seconds (fmt = 3)
convert(dm, 3)
#> Nelson's Column                           51°30′27.95″ (N/E)
#> Ostravice                                 49°32′46.36″ (N/E)
#> Tally Ho                                  48°06′26.03″ (N/E)
#> Washington Monument                       38°53′22.18″ (N/E)
#> Null Island                                0°00′00.00″ (N/E)
#> Tristan da Cunha                          37°06′42.26″ (S/W)
#> Mawson Peak                               53°06′17.21″ (S/W)
#> Silvio Pettirossi International Airport   25°14′24.56″ (S/W)
#> Nelson's Column                            0°07′40.52″ (S/W)
#> Ostravice                                 18°23′54.82″ (N/E)
#> Tally Ho                                 122°46′43.22″ (S/W)
#> Washington Monument                       77°02′06.87″ (S/W)
#> Null Island                                0°00′00.00″ (N/E)
#> Tristan da Cunha                          12°17′19.07″ (S/W)
#> Mawson Peak                               73°31′02.22″ (N/E)
#> Silvio Pettirossi International Airport   57°31′09.22″ (S/W)

## Convert to decimal degrees (fmt = 1)
convert(dm, 1)
#> Nelson's Column                            51.507765°
#> Ostravice                                  49.546210°
#> Tally Ho                                   48.107232°
#> Washington Monument                        38.889493°
#> Null Island                                 0.000000°
#> Tristan da Cunha                          -37.111740°
#> Mawson Peak                               -53.104782°
#> Silvio Pettirossi International Airport   -25.240155°
#> Nelson's Column                            -0.127923°
#> Ostravice                                  18.398562°
#> Tally Ho                                 -122.778672°
#> Washington Monument                       -77.035242°
#> Null Island                                 0.000000°
#> Tristan da Cunha                          -12.288630°
#> Mawson Peak                                73.517283°
#> Silvio Pettirossi International Airport   -57.519227°

###
## Continuing example from `waypoints()`...
## "waypoints" object in degrees, minutes and seconds

as_waypoints(wp, fmt = 3)
#>                                                Latitude        Longitude 
#>                                           ______________  _______________
#> Nelson's Column                           51°30′27.95″ N    0°07′40.53″ W
#> Ostravice                                 49°32′46.36″ N   18°23′54.82″ E
#> Tally Ho                                  48°06′26.04″ N  122°46′43.22″ W
#> Washington Monument                       38°53′22.18″ N   77°02′06.87″ W
#> Null Island                                0°00′00.00″ N    0°00′00.00″ E
#> Tristan da Cunha                          37°06′42.26″ S   12°17′19.07″ W
#> Mawson Peak                               53°06′17.21″ S   73°31′02.22″ E
#> Silvio Pettirossi International Airport   25°14′24.56″ S   57°31′09.21″ W

## Convert to degrees and minutes (fmt = 2)
convert(wp, 2)
#>                                               Latitude       Longitude 
#>                                           _____________  ______________
#> Nelson's Column                           51°30.4658′ N    0°07.6755′ W
#> Ostravice                                 49°32.7727′ N   18°23.9137′ E
#> Tally Ho                                  48°06.4340′ N  122°46.7203′ W
#> Washington Monument                       38°53.3697′ N   77°02.1145′ W
#> Null Island                                0°00.0000′ N    0°00.0000′ E
#> Tristan da Cunha                          37°06.7043′ S   12°17.3178′ W
#> Mawson Peak                               53°06.2868′ S   73°31.0370′ E
#> Silvio Pettirossi International Airport   25°14.4093′ S   57°31.1535′ W

## Convert to decimal degrees (fmt = 1)
convert(wp, 1)
#>                                             Latitude     Longitude 
#>                                           ___________  ____________
#> Nelson's Column                            51.507764°    -0.127925°
#> Ostravice                                  49.546211°    18.398561°
#> Tally Ho                                   48.107233°  -122.778672°
#> Washington Monument                        38.889494°   -77.035242°
#> Null Island                                 0.000000°     0.000000°
#> Tristan da Cunha                          -37.111739°   -12.288631°
#> Mawson Peak                               -53.104781°    73.517283°
#> Silvio Pettirossi International Airport   -25.240156°   -57.519225°

rm(dm, wp)