Geographic or GPS Coordinate Class

coords() creates a robust representation of a series of geographic or GPS coordinates instantiated as an object of class "coords".

coords() and replacement form function coords()<- also convert the format of existing objects of class "coords" between (i) decimal degrees, (ii) degrees and minutes, and (iii) degrees, minutes and seconds.

Usage

coords(nv, fmt = 1L)

coords(nv) <- value

# S3 method for class 'coords'
print(cd)

Arguments

nv

numeric vector of coordinate values, optionally named.

fmt, value

integer, 1L, 2L or 3L, indicating the current or desired coordinate format; default 1L.

cd

object of class "coords", as created by function coords().

Value

An object of class "coords" comprising a numeric vector with a boolean vector attribute "valid", indicating whether the individual coordinate values are indeed valid, as described under Details.

Details

Individual values provided in the numeric vector argument nv should have a decimal point after the number of whole degrees in the case of decimal degrees, after the number of whole minutes in the case of degrees and minutes, and after the number of whole seconds in the case of degrees, minutes and seconds.

The fmt argument is used to provide the format of values in a numeric vector to be converted into a "coords" object, and the desired format if a "coords" object is to be converted to a new format. fmt should be 1 for decimal degrees, 2 for degrees and minutes, and 3 for degrees, minutes and seconds. Note that following conversion, the original data structure is modified such that its values are as described in the previous paragraph, and may be inspected using standard R code, see examples.

The values of a newly created "coords" object are validated to ensure their being plausible geographic locations as described under validate(). Likewise, a check is made to ensure that that an existing "coords" object to be converted to a new format has already been validated; if not, it is re-validated.

See also

Examples

## Named numeric vector representing degrees and minutes
(num_dm <- c(
        5130.4659, 4932.7726, 4806.4339, 3853.3696, 0.0000, -3706.7044, -5306.2869, -5731.1536,
        -007.6754, 1823.9137, -12246.7203, -7702.1145,0.0000, -1217.3178, 7331.0370, -2514.4093
    ) |>
    setNames(rep(
        c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
          "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"), 2
    ))
)
#>                         Nelson's Column                               Ostravice 
#>                               5130.4659                               4932.7726 
#>                                Tally Ho                     Washington Monument 
#>                               4806.4339                               3853.3696 
#>                             Null Island                        Tristan da Cunha 
#>                                  0.0000                              -3706.7044 
#>                             Mawson Peak Silvio Pettirossi International Airport 
#>                              -5306.2869                              -5731.1536 
#>                         Nelson's Column                               Ostravice 
#>                                 -7.6754                               1823.9137 
#>                                Tally Ho                     Washington Monument 
#>                             -12246.7203                              -7702.1145 
#>                             Null Island                        Tristan da Cunha 
#>                                  0.0000                              -1217.3178 
#>                             Mawson Peak Silvio Pettirossi International Airport 
#>                               7331.0370                              -2514.4093 

## Create "coords" object of degrees and minutes
coords(num_dm, 2)
#>                         Nelson's Column   51°30.4659′ (N/E)
#>                               Ostravice   49°32.7726′ (N/E)
#>                                Tally Ho   48°06.4339′ (N/E)
#>                     Washington Monument   38°53.3696′ (N/E)
#>                             Null Island    0°00.0000′ (N/E)
#>                        Tristan da Cunha   37°06.7044′ (S/W)
#>                             Mawson Peak   53°06.2869′ (S/W)
#> Silvio Pettirossi International Airport   57°31.1536′ (S/W)
#>                         Nelson's Column    0°07.6754′ (S/W)
#>                               Ostravice   18°23.9137′ (N/E)
#>                                Tally Ho  122°46.7203′ (S/W)
#>                     Washington Monument   77°02.1145′ (S/W)
#>                             Null Island    0°00.0000′ (N/E)
#>                        Tristan da Cunha   12°17.3178′ (S/W)
#>                             Mawson Peak   73°31.0370′ (N/E)
#> Silvio Pettirossi International Airport   25°14.4093′ (S/W)

## Convert "coords" object to degrees, minutes and seconds
coords(num_dm) <- 3
num_dm
#>                         Nelson's Column   51°30′27.95″ (N/E)
#>                               Ostravice   49°32′46.36″ (N/E)
#>                                Tally Ho   48°06′26.03″ (N/E)
#>                     Washington Monument   38°53′22.18″ (N/E)
#>                             Null Island    0°00′00.00″ (N/E)
#>                        Tristan da Cunha   37°06′42.26″ (S/W)
#>                             Mawson Peak   53°06′17.21″ (S/W)
#> Silvio Pettirossi International Airport   57°31′09.22″ (S/W)
#>                         Nelson's Column    0°07′40.52″ (S/W)
#>                               Ostravice   18°23′54.82″ (N/E)
#>                                Tally Ho  122°46′43.22″ (S/W)
#>                     Washington Monument   77°02′06.87″ (S/W)
#>                             Null Island    0°00′00.00″ (N/E)
#>                        Tristan da Cunha   12°17′19.07″ (S/W)
#>                             Mawson Peak   73°31′02.22″ (N/E)
#> Silvio Pettirossi International Airport   25°14′24.56″ (S/W)

## Convert to decimal degrees
coords(num_dm) <- 1
num_dm
#>                         Nelson's Column    51.507765°
#>                               Ostravice    49.546210°
#>                                Tally Ho    48.107232°
#>                     Washington Monument    38.889493°
#>                             Null Island     0.000000°
#>                        Tristan da Cunha   -37.111740°
#>                             Mawson Peak   -53.104782°
#> Silvio Pettirossi International Airport   -57.519227°
#>                         Nelson's Column    -0.127923°
#>                               Ostravice    18.398562°
#>                                Tally Ho  -122.778672°
#>                     Washington Monument   -77.035242°
#>                             Null Island     0.000000°
#>                        Tristan da Cunha   -12.288630°
#>                             Mawson Peak    73.517283°
#> Silvio Pettirossi International Airport   -25.240155°

## Convert back to degrees and minutes
coords(num_dm) <- 2
num_dm
#>                         Nelson's Column   51°30.4659′ (N/E)
#>                               Ostravice   49°32.7726′ (N/E)
#>                                Tally Ho   48°06.4339′ (N/E)
#>                     Washington Monument   38°53.3696′ (N/E)
#>                             Null Island    0°00.0000′ (N/E)
#>                        Tristan da Cunha   37°06.7044′ (S/W)
#>                             Mawson Peak   53°06.2869′ (S/W)
#> Silvio Pettirossi International Airport   57°31.1536′ (S/W)
#>                         Nelson's Column    0°07.6754′ (S/W)
#>                               Ostravice   18°23.9137′ (N/E)
#>                                Tally Ho  122°46.7203′ (S/W)
#>                     Washington Monument   77°02.1145′ (S/W)
#>                             Null Island    0°00.0000′ (N/E)
#>                        Tristan da Cunha   12°17.3178′ (S/W)
#>                             Mawson Peak   73°31.0370′ (N/E)
#> Silvio Pettirossi International Airport   25°14.4093′ (S/W)

## Convert to decimal degrees and show as a numeric vector
coords(num_dm) <- 1
as.numeric(num_dm)
#>  [1]   51.5077650   49.5462100   48.1072317   38.8894933    0.0000000
#>  [6]  -37.1117400  -53.1047817  -57.5192267   -0.1279233   18.3985617
#> [11] -122.7786717  -77.0352417    0.0000000  -12.2886300   73.5172833
#> [16]  -25.2401550

rm(num_dm)