Geographic or GPS Coordinate Class

coords() creates a robust representation of a series of geographic or GPS coordinates instantiated as an object of class "coords".

coords() and replacement form coords()<- also convert the format of existing objects of class "coords" between (i) decimal degrees, (ii) degrees and minutes, and (iii) degrees, minutes and seconds.

Usage

coords(nv, fmt = 1L)

coords(nv) <- value

# S3 method for class 'coords'
format(x, ...)

# S3 method for class 'coords'
print(x, ...)

Arguments

nv

numeric vector of coordinate values, optionally named.

fmt, value

integer, 1L, 2L or 3L, indicating the current or desired coordinate format; default 1L.

x

object of class "coords" created by function coords().

...

further arguments passed to or from other methods.

Value

An object of class "coords", comprising the original a numeric vector argument nv with values possibly converted as appropriate and additional attributes: –

"fmt"

the coordinate format.

"valid"

a logical vector indicating whether individual coordinate values are valid geographic locations.

Details

Individual values provided in the numeric vector argument nv should have a decimal point after the number of whole degrees in the case of decimal degrees, after the number of whole minutes in the case of degrees and minutes, and after the number of whole seconds in the case of degrees, minutes and seconds.

The fmt argument should be 1L to represent decimal degrees, 2L for degrees and minutes, and 3L for degrees, minutes and seconds and is used to provide both the format of values in numeric vector argument nv to be converted into a "coords" object and the desired format if a "coords" object is to be converted to a new format. Note that on conversion of a "coords" object, the original numeric vector argument nv is modified such that the values are as described in the previous paragraph, and may be inspected using standard R code, see examples.

The values of a newly created "coords" object are checked to ensure they are valid geographic locations as described under validate(). Likewise, a check is made to ensure that an existing "coords" object to be converted to a new format has already been validated; if not, it is re-validated.

See also

Examples

## Numeric vector representing degrees and minutes
dm <- c(5130.4659, 4932.7726, 4806.4339, 3853.3696, 0.0000, -3706.7044, -5306.2869, -2514.4093,
        -007.6754, 1823.9137, -12246.7203, -7702.1145, 0.0000, -1217.3178, 7331.0370, -5731.1536)

## Create a "coords" object of degrees and minutes
coords(dm, 2)
#>  51°30.4659′ (N/E)
#>  49°32.7726′ (N/E)
#>  48°06.4339′ (N/E)
#>  38°53.3696′ (N/E)
#>   0°00.0000′ (N/E)
#>  37°06.7044′ (S/W)
#>  53°06.2869′ (S/W)
#>  25°14.4093′ (S/W)
#>   0°07.6754′ (S/W)
#>  18°23.9137′ (N/E)
#> 122°46.7203′ (S/W)
#>  77°02.1145′ (S/W)
#>   0°00.0000′ (N/E)
#>  12°17.3178′ (S/W)
#>  73°31.0370′ (N/E)
#>  57°31.1536′ (S/W)

## Name the "coords" object
names(dm) <- rep(c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
                   "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"), 2)
dm
#>                         Nelson's Column   51°30.4659′ (N/E)
#>                               Ostravice   49°32.7726′ (N/E)
#>                                Tally Ho   48°06.4339′ (N/E)
#>                     Washington Monument   38°53.3696′ (N/E)
#>                             Null Island    0°00.0000′ (N/E)
#>                        Tristan da Cunha   37°06.7044′ (S/W)
#>                             Mawson Peak   53°06.2869′ (S/W)
#> Silvio Pettirossi International Airport   25°14.4093′ (S/W)
#>                         Nelson's Column    0°07.6754′ (S/W)
#>                               Ostravice   18°23.9137′ (N/E)
#>                                Tally Ho  122°46.7203′ (S/W)
#>                     Washington Monument   77°02.1145′ (S/W)
#>                             Null Island    0°00.0000′ (N/E)
#>                        Tristan da Cunha   12°17.3178′ (S/W)
#>                             Mawson Peak   73°31.0370′ (N/E)
#> Silvio Pettirossi International Airport   57°31.1536′ (S/W)

## Convert to degrees, minutes and seconds
coords(dm) <- 3
dm
#>                         Nelson's Column   51°30′27.95″ (N/E)
#>                               Ostravice   49°32′46.36″ (N/E)
#>                                Tally Ho   48°06′26.03″ (N/E)
#>                     Washington Monument   38°53′22.18″ (N/E)
#>                             Null Island    0°00′00.00″ (N/E)
#>                        Tristan da Cunha   37°06′42.26″ (S/W)
#>                             Mawson Peak   53°06′17.21″ (S/W)
#> Silvio Pettirossi International Airport   25°14′24.56″ (S/W)
#>                         Nelson's Column    0°07′40.52″ (S/W)
#>                               Ostravice   18°23′54.82″ (N/E)
#>                                Tally Ho  122°46′43.22″ (S/W)
#>                     Washington Monument   77°02′06.87″ (S/W)
#>                             Null Island    0°00′00.00″ (N/E)
#>                        Tristan da Cunha   12°17′19.07″ (S/W)
#>                             Mawson Peak   73°31′02.22″ (N/E)
#> Silvio Pettirossi International Airport   57°31′09.22″ (S/W)

## Convert to decimal degrees
coords(dm) <- 1
dm
#>                         Nelson's Column    51.507765°
#>                               Ostravice    49.546210°
#>                                Tally Ho    48.107232°
#>                     Washington Monument    38.889493°
#>                             Null Island     0.000000°
#>                        Tristan da Cunha   -37.111740°
#>                             Mawson Peak   -53.104782°
#> Silvio Pettirossi International Airport   -25.240155°
#>                         Nelson's Column    -0.127923°
#>                               Ostravice    18.398562°
#>                                Tally Ho  -122.778672°
#>                     Washington Monument   -77.035242°
#>                             Null Island     0.000000°
#>                        Tristan da Cunha   -12.288630°
#>                             Mawson Peak    73.517283°
#> Silvio Pettirossi International Airport   -57.519227°

## Decimal degrees as a standard numeric vector
as.numeric(dm)
#>  [1]   51.5077650   49.5462100   48.1072317   38.8894933    0.0000000
#>  [6]  -37.1117400  -53.1047817  -25.2401550   -0.1279233   18.3985617
#> [11] -122.7786717  -77.0352417    0.0000000  -12.2886300   73.5172833
#> [16]  -57.5192267

## Convert to degrees and minutes, and format as a character vector
coords(dm) <- 2
(dm_chr <- format(dm))
#>  [1] " 51°30.4659′ (N/E)" " 49°32.7726′ (N/E)" " 48°06.4339′ (N/E)"
#>  [4] " 38°53.3696′ (N/E)" "  0°00.0000′ (N/E)" " 37°06.7044′ (S/W)"
#>  [7] " 53°06.2869′ (S/W)" " 25°14.4093′ (S/W)" "  0°07.6754′ (S/W)"
#> [10] " 18°23.9137′ (N/E)" "122°46.7203′ (S/W)" " 77°02.1145′ (S/W)"
#> [13] "  0°00.0000′ (N/E)" " 12°17.3178′ (S/W)" " 73°31.0370′ (N/E)"
#> [16] " 57°31.1536′ (S/W)"

## Output using {base} cat()
cat(dm_chr, fill = 18, labels = paste0("{#", 1:16, "}:"))
#> {#1}:  51°30.4659′ (N/E) 
#> {#2}:  49°32.7726′ (N/E) 
#> {#3}:  48°06.4339′ (N/E) 
#> {#4}:  38°53.3696′ (N/E) 
#> {#5}:   0°00.0000′ (N/E) 
#> {#6}:  37°06.7044′ (S/W) 
#> {#7}:  53°06.2869′ (S/W) 
#> {#8}:  25°14.4093′ (S/W) 
#> {#9}:   0°07.6754′ (S/W) 
#> {#10}:  18°23.9137′ (N/E) 
#> {#11}: 122°46.7203′ (S/W) 
#> {#12}:  77°02.1145′ (S/W) 
#> {#13}:   0°00.0000′ (N/E) 
#> {#14}:  12°17.3178′ (S/W) 
#> {#15}:  73°31.0370′ (N/E) 
#> {#16}:  57°31.1536′ (S/W)

rm(dm, dm_chr)