Format Coords or Waypoints

Format and print objects of class "coords" or "waypoints".

Usage

# S3 method for class 'coords'
print(x, ...)

# S3 method for class 'waypoints'
print(x, ...)

# S3 method for class 'coords'
format(x, ..., usenames = TRUE)

# S3 method for class 'waypoints'
format(x, ..., usenames = TRUE)

ll_headers(aswidth, fmt)

Arguments

x

object of class "coords" created by function as_coords(), or class "waypoints" created by function as_waypoints().

...

further arguments passed to or from other methods.

usenames

logical, whether or not to include names in formatted output; default TRUE.

aswidth

character vector, used to match width of headers to formatted output.

fmt

integer, 1L, 2L or 3L, specifying the required coordinate format.

Value

The format() methods for both classes "coords" and "waypoints" return a character vector, respectively of length length(x) or nrow(x), and containing values formatted in decimal degrees, degrees and minutes, or degrees, minutes and seconds as appropriate.

Details

The format() methods for "coords" and "waypoints" objects output elegantly formatted character vector representations of their arguments, which are used by their respective print() methods.

Prior to formatting and printing, "coords" or "waypoints" objects are checked to ensure that their "valid" attribute (in the case of a "coords" object), or "validlat" and "validlon" attributes (in the case of a "waypoints" object) are present and all TRUE i.e., valid. If these attributes are found to contain any FALSE i.e. invalid values, a warning is issued and similarly, if these attributes are missing, a warning is issued and the objects are re-validated as described under validate().

ll_headers() outputs the headings "Latitude ... Longitude" formatted to the same width as argument aswidth, adjusted for format fmt and is primarily intended for use by the print() method for class "waypoints".

See also

Examples

## Continuing example from `coords()`...
## Named "coords" object in degrees and minutes

## Silently using print() method for class "coords"
as_coords(dm, fmt = 2)
#> Nelson's Column                           51°30.4659′ (N/E)
#> Ostravice                                 49°32.7726′ (N/E)
#> Tally Ho                                  48°06.4339′ (N/E)
#> Washington Monument                       38°53.3696′ (N/E)
#> Null Island                                0°00.0000′ (N/E)
#> Tristan da Cunha                          37°06.7044′ (S/W)
#> Mawson Peak                               53°06.2869′ (S/W)
#> Silvio Pettirossi International Airport   25°14.4093′ (S/W)
#> Nelson's Column                            0°07.6754′ (S/W)
#> Ostravice                                 18°23.9137′ (N/E)
#> Tally Ho                                 122°46.7203′ (S/W)
#> Washington Monument                       77°02.1145′ (S/W)
#> Null Island                                0°00.0000′ (N/E)
#> Tristan da Cunha                          12°17.3178′ (S/W)
#> Mawson Peak                               73°31.0370′ (N/E)
#> Silvio Pettirossi International Airport   57°31.1536′ (S/W)

## Format as a fixed-width character vector with names...
format(dm)
#>  [1] "Nelson's Column                           51°30.4659′ (N/E)"
#>  [2] "Ostravice                                 49°32.7726′ (N/E)"
#>  [3] "Tally Ho                                  48°06.4339′ (N/E)"
#>  [4] "Washington Monument                       38°53.3696′ (N/E)"
#>  [5] "Null Island                                0°00.0000′ (N/E)"
#>  [6] "Tristan da Cunha                          37°06.7044′ (S/W)"
#>  [7] "Mawson Peak                               53°06.2869′ (S/W)"
#>  [8] "Silvio Pettirossi International Airport   25°14.4093′ (S/W)"
#>  [9] "Nelson's Column                            0°07.6754′ (S/W)"
#> [10] "Ostravice                                 18°23.9137′ (N/E)"
#> [11] "Tally Ho                                 122°46.7203′ (S/W)"
#> [12] "Washington Monument                       77°02.1145′ (S/W)"
#> [13] "Null Island                                0°00.0000′ (N/E)"
#> [14] "Tristan da Cunha                          12°17.3178′ (S/W)"
#> [15] "Mawson Peak                               73°31.0370′ (N/E)"
#> [16] "Silvio Pettirossi International Airport   57°31.1536′ (S/W)"

## ...or without them
format(dm, usenames = FALSE)
#>  [1] " 51°30.4659′ (N/E)" " 49°32.7726′ (N/E)" " 48°06.4339′ (N/E)"
#>  [4] " 38°53.3696′ (N/E)" "  0°00.0000′ (N/E)" " 37°06.7044′ (S/W)"
#>  [7] " 53°06.2869′ (S/W)" " 25°14.4093′ (S/W)" "  0°07.6754′ (S/W)"
#> [10] " 18°23.9137′ (N/E)" "122°46.7203′ (S/W)" " 77°02.1145′ (S/W)"
#> [13] "  0°00.0000′ (N/E)" " 12°17.3178′ (S/W)" " 73°31.0370′ (N/E)"
#> [16] " 57°31.1536′ (S/W)"

###
## Continuing example from `waypoints()`...
## "waypoints" object in degrees, minutes and seconds

## Silently using print() method for class "waypoints"
as_waypoints(wp, fmt = 3)
#>                                                Latitude        Longitude 
#>                                           ______________  _______________
#> Nelson's Column                           51°30′27.95″ N    0°07′40.53″ W
#> Ostravice                                 49°32′46.36″ N   18°23′54.82″ E
#> Tally Ho                                  48°06′26.04″ N  122°46′43.22″ W
#> Washington Monument                       38°53′22.18″ N   77°02′06.87″ W
#> Null Island                                0°00′00.00″ N    0°00′00.00″ E
#> Tristan da Cunha                          37°06′42.26″ S   12°17′19.07″ W
#> Mawson Peak                               53°06′17.21″ S   73°31′02.22″ E
#> Silvio Pettirossi International Airport   25°14′24.56″ S   57°31′09.21″ W

## Format as a fixed-width character vector with names...
format(wp)
#> [1] "Nelson's Column                           51°30′27.95″ N    0°07′40.53″ W"
#> [2] "Ostravice                                 49°32′46.36″ N   18°23′54.82″ E"
#> [3] "Tally Ho                                  48°06′26.04″ N  122°46′43.22″ W"
#> [4] "Washington Monument                       38°53′22.18″ N   77°02′06.87″ W"
#> [5] "Null Island                                0°00′00.00″ N    0°00′00.00″ E"
#> [6] "Tristan da Cunha                          37°06′42.26″ S   12°17′19.07″ W"
#> [7] "Mawson Peak                               53°06′17.21″ S   73°31′02.22″ E"
#> [8] "Silvio Pettirossi International Airport   25°14′24.56″ S   57°31′09.21″ W"

## ...or without them
format(wp, usenames = FALSE)
#> [1] " 51°30′27.95″ N    0°07′40.53″ W" " 49°32′46.36″ N   18°23′54.82″ E"
#> [3] " 48°06′26.04″ N  122°46′43.22″ W" " 38°53′22.18″ N   77°02′06.87″ W"
#> [5] "  0°00′00.00″ N    0°00′00.00″ E" " 37°06′42.26″ S   12°17′19.07″ W"
#> [7] " 53°06′17.21″ S   73°31′02.22″ E" " 25°14′24.56″ S   57°31′09.21″ W"

rm(dm, wp)