Validate Coords or Waypoints

validate() validate objects of class "coords" or "waypoints".

Usage

# S3 method for class 'coords'
validate(cd)

# S3 method for class 'waypoints'
validate(df)

Arguments

cd

object of class "coords".

df

object of class "waypoints".

Value

validate() returns its argument with logical vector attribute "valid", or attributes "validlat" and "validlon" updated as appropriate for "coords" and' "waypoints" objects respectively.

Details

Individual coordinate values within "coords" or "waypoints" objects are validated to ensure their being plausible geographic locations.

To be valid, the absolute values of coordinates in degrees must not exceed 180, or 90 if degrees of latitude and, similarly, the absolute values of the minutes and seconds components, where given, must not exceed 60 degrees. Otherwise a warning will be issued and the "valid" attribute in the case of a "coords" object, or "validlat" and "validlon" attributes in the case of a "waypoints" object will be set to FALSE for any non-compliant coordinate values.

See also

Examples

## Continuing example from `coords()`, named numeric vector representing degrees and minutes

## Create "coords" object of degrees and minutes
coords(dm) <- 2

validate(dm)
#>                         Nelson's Column   51°30.4659′ (N/E)
#>                               Ostravice   49°32.7726′ (N/E)
#>                                Tally Ho   48°06.4339′ (N/E)
#>                     Washington Monument   38°53.3696′ (N/E)
#>                             Null Island    0°00.0000′ (N/E)
#>                        Tristan da Cunha   37°06.7044′ (S/W)
#>                             Mawson Peak   53°06.2869′ (S/W)
#> Silvio Pettirossi International Airport   25°14.4093′ (S/W)
#>                         Nelson's Column    0°07.6754′ (S/W)
#>                               Ostravice   18°23.9137′ (N/E)
#>                                Tally Ho  122°46.7203′ (S/W)
#>                     Washington Monument   77°02.1145′ (S/W)
#>                             Null Island    0°00.0000′ (N/E)
#>                        Tristan da Cunha   12°17.3178′ (S/W)
#>                             Mawson Peak   73°31.0370′ (N/E)
#> Silvio Pettirossi International Airport   57°31.1536′ (S/W)

## Change first value to have more than 60 minutes
dm[1] <- 5160.4659

validate(dm)
#> Warning: Validation failed!
#> Warning: Printing invalid coords!
#>                         Nelson's Column   51°60.4659′ (N/E)
#>                               Ostravice   49°32.7726′ (N/E)
#>                                Tally Ho   48°06.4339′ (N/E)
#>                     Washington Monument   38°53.3696′ (N/E)
#>                             Null Island    0°00.0000′ (N/E)
#>                        Tristan da Cunha   37°06.7044′ (S/W)
#>                             Mawson Peak   53°06.2869′ (S/W)
#> Silvio Pettirossi International Airport   25°14.4093′ (S/W)
#>                         Nelson's Column    0°07.6754′ (S/W)
#>                               Ostravice   18°23.9137′ (N/E)
#>                                Tally Ho  122°46.7203′ (S/W)
#>                     Washington Monument   77°02.1145′ (S/W)
#>                             Null Island    0°00.0000′ (N/E)
#>                        Tristan da Cunha   12°17.3178′ (S/W)
#>                             Mawson Peak   73°31.0370′ (N/E)
#> Silvio Pettirossi International Airport   57°31.1536′ (S/W)

## Examine valid attribute of dm
attr(dm, "valid")
#>  [1] FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#> [13]  TRUE  TRUE  TRUE  TRUE

###
## Continuing example from `waypoints()`, data frame representing waypoint names and latitude
## and longitude values in decimal degrees

## Create "waypoints" object of decimal degrees
waypoints(wp1) <- 1

validate(wp1)
#>                                             Latitude     Longitude 
#>                                           ___________  ____________
#>                         Nelson's Column    51.507765°    -0.127924°
#>                               Ostravice    49.546210°    18.398562°
#>                                Tally Ho    48.107232°  -122.778671°
#>                     Washington Monument    38.889494°   -77.035242°
#>                             Null Island     0.000000°     0.000000°
#>                        Tristan da Cunha   -37.111740°   -12.288630°
#>                             Mawson Peak   -53.104781°    73.517283°
#> Silvio Pettirossi International Airport   -25.240156°   -57.519227°

## Change penultimate latitude absolute value to have more than 90 degrees
wp1$lat[7] <- -93.104781

validate(wp1)
#> Warning: Validation of latitude failed!
#> Warning: Invalid latitude!
#>                                             Latitude     Longitude 
#>                                           ___________  ____________
#>                         Nelson's Column    51.507765°    -0.127924°
#>                               Ostravice    49.546210°    18.398562°
#>                                Tally Ho    48.107232°  -122.778671°
#>                     Washington Monument    38.889494°   -77.035242°
#>                             Null Island     0.000000°     0.000000°
#>                        Tristan da Cunha   -37.111740°   -12.288630°
#>                             Mawson Peak   -93.104781°    73.517283°
#> Silvio Pettirossi International Airport   -25.240156°   -57.519227°

## Examine validlat attribute of wp1
attr(wp1, "validlat")
#> [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE

rm(dm, wp1)