Geographic or GPS Waypoint Class

waypoints() creates a robust representation of a series of geographic or GPS waypoints instantiated as an object of class "waypoints".

waypoints() and replacement form function waypoints()<- also convert the format of existing objects of class "waypoints" between (i) decimal degrees, (ii) degrees and minutes, and (iii) degrees, minutes and seconds.

Usage

waypoints(df, fmt = 1L)

waypoints(df) <- value

# S3 method for class 'waypoints'
print(df)

Arguments

df

a dataframe containing at least two numeric columns containing corresponding values of latitude and longitude in each row representing a waypoint.

fmt, value

integer, 1, 2 or 3, indicating the current or desired coordinate format; default 1.

Value

An object of class "waypoints" comprising a data.frame with two boolean vector attributes "validlat" and "validlon" indicating whether the individual coordinate values are indeed valid, as described under Details.

Details

Individual values provided in the numeric vector latitude and longitude columns of argument df should have a decimal point after the number of whole degrees in the case of decimal degrees, after the number of whole minutes in the case of degrees and minutes, and after the number of whole seconds in the case of degrees, minutes and seconds.

The fmt argument is used to provide the format of values in a dataframe to be converted into a "waypoints" object, and the desired format if a "waypoints" object is to be converted to a new format. fmt should be 1 for decimal degrees, 2 for degrees and minutes, and 3 for degrees, minutes and seconds. Note that following conversion, the original data structure is modified such that the latitude and longitude values are as described in the previous paragraph, and may be inspected using standard R code, see examples.

The latitude and longitude values of a newly created "waypoints" object are validated to ensure their being plausible geographic locations as described under validate(). Likewise, a check is made to ensure that an existing "coords" object to be converted to a new format has already been validated; if not, it is re-validated.

See also

Examples

# Dataframe representing waypoint names and latitude and longitude values in decimal degrees
(wp1 <- data.frame(
    name = c("Nelson's Column", "Ostravice", "Tally Ho", "Washington Monument", "Null Island",
             "Tristan da Cunha", "Mawson Peak", "Silvio Pettirossi International Airport"),
    lat = c(51.507765, 49.54621, 48.107232, 38.889494, 0, -37.11174, -53.104781, -57.519227),
    lon = c(-0.127924, 18.398562, -122.778671, -77.035242, 0, -12.28863, 73.517283, -25.240156)
))
#>                                      name       lat         lon
#> 1                         Nelson's Column  51.50776   -0.127924
#> 2                               Ostravice  49.54621   18.398562
#> 3                                Tally Ho  48.10723 -122.778671
#> 4                     Washington Monument  38.88949  -77.035242
#> 5                             Null Island   0.00000    0.000000
#> 6                        Tristan da Cunha -37.11174  -12.288630
#> 7                             Mawson Peak -53.10478   73.517283
#> 8 Silvio Pettirossi International Airport -57.51923  -25.240156

# Create "waypoints" object of decimal degrees
waypoints(wp1)
#>  Latitude     Longitude
#>  ___________  ____________
#>   51.507765°    -0.127924°  Nelson's Column
#>   49.546210°    18.398562°  Ostravice
#>   48.107232°  -122.778671°  Tally Ho
#>   38.889494°   -77.035242°  Washington Monument
#>    0.000000°     0.000000°  Null Island
#>  -37.111740°   -12.288630°  Tristan da Cunha
#>  -53.104781°    73.517283°  Mawson Peak
#>  -57.519227°   -25.240156°  Silvio Pettirossi International Airport

## Convert "waypoints" object to degrees and minutes
waypoints(wp1) <- 2
wp1
#>  Latitude       Longitude
#>  _____________  ______________
#>  51°30.4659' N    0°07.6754' W  Nelson's Column
#>  49°32.7726' N   18°23.9137' E  Ostravice
#>  48°06.4339' N  122°46.7203' W  Tally Ho
#>  38°53.3696' N   77°02.1145' W  Washington Monument
#>   0°00.0000' N    0°00.0000' E  Null Island
#>  37°06.7044' S   12°17.3178' W  Tristan da Cunha
#>  53°06.2869' S   73°31.0370' E  Mawson Peak
#>  57°31.1536' S   25°14.4094' W  Silvio Pettirossi International Airport

## Convert "waypoints" object to degrees, minutes and seconds
waypoints(wp1) <- 3
wp1
#>  Latitude        Longitude
#>  ______________  _______________
#>  51°30'27.95" N    0°07'40.53" W  Nelson's Column
#>  49°32'46.36" N   18°23'54.82" E  Ostravice
#>  48°06'26.04" N  122°46'43.22" W  Tally Ho
#>  38°53'22.18" N   77°02'06.87" W  Washington Monument
#>   0°00'00.00" N    0°00'00.00" E  Null Island
#>  37°06'42.26" S   12°17'19.07" W  Tristan da Cunha
#>  53°06'17.21" S   73°31'02.22" E  Mawson Peak
#>  57°31'09.22" S   25°14'24.56" W  Silvio Pettirossi International Airport

## Convert back to decimal degrees
waypoints(wp1) <- 1
wp1
#>  Latitude     Longitude
#>  ___________  ____________
#>   51.507765°    -0.127924°  Nelson's Column
#>   49.546210°    18.398562°  Ostravice
#>   48.107232°  -122.778671°  Tally Ho
#>   38.889494°   -77.035242°  Washington Monument
#>    0.000000°     0.000000°  Null Island
#>  -37.111740°   -12.288630°  Tristan da Cunha
#>  -53.104781°    73.517283°  Mawson Peak
#>  -57.519227°   -25.240156°  Silvio Pettirossi International Airport

## Convert back to degrees, minutes and seconds and to a data.frame
waypoints(wp1) <- 3
class(wp1) <- "data.frame"
wp1
#>                                      name       lat           lon
#> 1                         Nelson's Column  513028.0     -740.5264
#> 2                               Ostravice  493246.4   182354.8232
#> 3                                Tally Ho  480626.0 -1224643.2156
#> 4                     Washington Monument  385322.2  -770206.8712
#> 5                             Null Island       0.0        0.0000
#> 6                        Tristan da Cunha -370642.3  -121719.0680
#> 7                             Mawson Peak -530617.2   733102.2188
#> 8 Silvio Pettirossi International Airport -573109.2  -251424.5616

rm(wp1)